refactor : 승,무,패 enum 공통으로 사용
