+++ src/main/java/blackjack/domain/BlackjackGame.java
@@ -1,6 +1,7 @@
 package blackjack.domain;
 
 import blackjack.util.CardPickerGenerator;
+import blackjack.util.WinningResult;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
