--- src/main/java/blackjack/domain/WinningResult.java
@@ -1,9 +1,0 @@
-package blackjack.domain;
-
-public enum WinningResult {
-    WIN,
-    LOSE,
-    PUSH,
-    ;
-
-}
