+++ src/test/java/blackjack/domain/BlackjackGameTest.java
@@ -3,6 +3,7 @@ package blackjack.domain;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
 import blackjack.util.CardPickerGenerator;
+import blackjack.util.WinningResult;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
